requirements.txt
Flask 
Flask-Script
Flask-Bootstrap
mysql-connector